# Conflicts:
#	custom_components/peaqev/peaqservice/chargecontroller/ichargecontroller.py
#	custom_components/peaqev/peaqservice/hub/sensors/ihub_sensors.py

+++ custom_components/peaqev/peaqservice/chargecontroller/charger/chargermodel.py
@@ -6,6 +6,6 @@ class ChargerModel:
     running: bool = False
     disable_current_updates: bool = False
     session_active: bool = False
-    latest_charger_call: int = 0
+    latest_charger_call: float = 0
     lastest_call_off: float = 0
     charger_state_mismatch: bool = False
